Direct stderr to /dev/null in wal-prefetch forks
Without this change DaemonContext will close stderr for the forked
processes and that results in `gpg` printing a warning to stdout that
it had to reopen stderr. The warning will make lzop fail since the
input stream is not a valid archive anymore.

When `lzop` crashes it causes the prefetch process to get stuck since
the buffers between the processes are not properly closed and drained.